Almost to an even better Sakura Burst
I kind of screwed up on some of the code. It does work as it should, but I used it the wrong way. I should of actually created my own emitter on the flower object and used that emitter instead, or since the bullet has a spell library on it, I should be referring the newly spawn rotation emitters to their actual position, which I think that's where the problem is.