Smaller Magic Circle Orbit
I have the Smaller Magic Circles for Raven to rotate around the bigger Magic Circle.
Now, I need to fix the directional shooting, because it's going directly with the rotation of the magic circle (this only applies for the Smaller Magic Circle instead of the big one for some reason).