Added SpriteBank/Data
Adding a SpriteBank to each CharacterProfile, and each SpriteBank will contain sprite animations that can be accessed by name or by index (which by index is the easiest approach). The Stats and CharacterBank of the character will be sent to the GameManager to carry that data persistently. I've also added a StageMap class, as well as a Stage class that hold 5 different StageMap, each one for each difficulty. I have passed a DifficultyIndex into the gameManager, and have the Stage load will the corresponding map depending on this index (which works).

Next, I will redo the calculations for the Stats, as well as put in newly done sprites for Raven, Maple, and now Hanabi.