Added New Asset: BulletPro
Added a new asset to speed up production for the game. This will allow me to add more complexity to character spells. The rendering in bullets is also a huge improvement, because I need a lot of bullets to be on the screen (and I mean a lot).

We have also made it so that if a stat is at or above 30, it's given a 5 star rating. For the in-game logic, I well alter these values with some creating equations to add balance to the game.

Added a new static function to the EventManager called Watch, in which checks for a given condition to be true before executing.