New Gameplay Layout
Altered the Gameplay GUI Interface to maximize player experience. We removed the Spell Library UI, and instead will be implemented into the pause menu, which has you select an option where you can check a character's spell library.

Also, we added in another UI that'll be part of our Power Grade System, which is a mechanic commonly known in a Danmaku/Shoot'em Up game.