Created Power System
I added in a Power Grade System that'll make the character's Witch Sign more and more powerful as they collect more magic essence. As of right now, the default level that the player will be starting at will be at level 3. The power system uses the modulus operator for setting the meter fill, as well as the level that the player is at. This makes it so the meter can be added and subtracted accurately whenever Power Levels change.