SequenceEventScript.cs Created
With a set format, I can create a list of events to happen at certain intervals. Currently, we got the system working. Now we just have to implement each part of each preBuilt events that you can call using the system. First, we'll see if we can get a stage animation implements, and have it played whenever we want it to.